refactor: move web_search provider/API-key UI into ToolConfigModal
The Web Search provider dropdown and API-key picker previously lived in
Settings → General → Web Search, forcing users to jump out of the tool
configuration flow to point the tool at a provider.

Move both inputs into ToolConfigModal's web_search panel next to Max
Results, and remove the Web Search section from GeneralSettings. The
underlying state stays on PluginData (webSearchProvider,
webSearchApiKeyId) — provider/key remain global rather than per-agent
because the API key is a secret rotated in one place, mirroring how
provider credentials are stored elsewhere in the plugin.

Also updates the web_search tool card description in AgentEditorModal
to point users at the tool's Configure panel instead of General
settings.
